Refactoring, graceful exit
